Move repeated dashboard code to Dashboard component
