+++ src/dashboard/Sidebar.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import clsx from 'clsx';
+import Drawer from '@material-ui/core/Drawer';
+import List from '@material-ui/core/List';
+import Divider from '@material-ui/core/Divider';
+import IconButton from '@material-ui/core/IconButton';
+import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
+import { mainListItems } from './listItems';
+
+export default function Sidebar({classes, open, handleDrawerClose}) {
+  return (
+    <Drawer
+    variant="permanent"
+    classes={{
+      paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose),
+    }}
+    open={open}
+    >
+    <div className={classes.toolbarIcon}>
+      <IconButton onClick={handleDrawerClose}>
+      <ChevronLeftIcon />
+      </IconButton>
+    </div>
+    <Divider />
+    <List>{mainListItems}</List>
+    </Drawer>
+  );
+}
